+++ module/herosystem6e.mjs
@@ -590,7 +590,7 @@ Hooks.on("updateWorldTime", async (worldTime, options) => {
             }
 
             // Active Effects
-            if (!actor.inCombat) {
+            if (!actor.inCombat || actor.temporaryEffects.find((o) => o.name === "TakeRecovery")) {
                 await expireEffects(actor);
             }