+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Implemented SIMPLIFIED healing. [#1256](https://github.com/dmdorman/hero6e-foundryvtt/issues/1407)
 - You can now select the override key you want to use. It used to be hard coded to be SHIFT. Changed the default to be ControlLeft because SHIFT movement was snapless and thus awkward. [#1410](https://github.com/dmdorman/hero6e-foundryvtt/issues/1410)
 - Fixed missing CSL from AOE to-hit template placement. [#1392](https://github.com/dmdorman/hero6e-foundryvtt/issues/1392)
+- Taking a Recovery applies 1/2 DCV penalty for 1 second (until start of next segment). [#1405](https://github.com/dmdorman/hero6e-foundryvtt/issues/1405)
 
 ## Version 4.0.3